k/group: moved all group metadata types to old namespace
Moved all group metadata related types to `kafka::old` namespace. This
way we will be able to translate between old and new representations.
