+++ src/v/kafka/server/group_stm.cc
@@ -5,7 +5,7 @@
 
 namespace kafka {
 
-void group_stm::overwrite_metadata(group_log_group_metadata&& metadata) {
+void group_stm::overwrite_metadata(old::group_log_group_metadata&& metadata) {
     _metadata = std::move(metadata);
     _is_loaded = true;
 }
@@ -17,7 +17,7 @@ void group_stm::remove_offset(const model::topic_partition& key) {
 void group_stm::update_offset(
   const model::topic_partition& key,
   model::offset offset,
-  group_log_offset_metadata&& meta) {
+  old::group_log_offset_metadata&& meta) {
     _offsets[key] = logged_metadata{
       .log_offset = offset, .metadata = std::move(meta)};
 }
@@ -72,7 +72,7 @@ void group_stm::commit(model::producer_identity pid) {
     }
 
     for (const auto& [tp, md] : prepared_it->second.offsets) {
-        group_log_offset_metadata val{
+        old::group_log_offset_metadata val{
           .offset = md.offset,
           .leader_epoch = 0, // we never use leader_epoch down the stack
           .metadata = md.metadata};
@@ -93,22 +93,6 @@ void group_stm::abort(
         return;
     }
     _prepared_txs.erase(prepared_it);
-}
-
-std::ostream& operator<<(std::ostream& os, const group_log_offset_key& key) {
-    fmt::print(
-      os,
-      "group {} topic {} partition {}",
-      key.group(),
-      key.topic(),
-      key.partition());
-    return os;
-}
-
-std::ostream&
-operator<<(std::ostream& os, const group_log_offset_metadata& md) {
-    fmt::print(os, "offset {}", md.offset());
-    return os;
 }
 
 } // namespace kafka
