+++ src/v/kafka/server/group_stm.h
@@ -3,6 +3,7 @@
 #include "cluster/partition.h"
 #include "kafka/protocol/fwd.h"
 #include "kafka/server/group.h"
+#include "kafka/server/group_stm.h"
 #include "kafka/server/member.h"
 #include "kafka/types.h"
 #include "model/fundamental.h"
@@ -22,43 +23,6 @@
 #include <vector>
 
 namespace kafka {
-
-/**
- * the value type of a group metadata log record.
- */
-struct group_log_group_metadata {
-    kafka::protocol_type protocol_type;
-    kafka::generation_id generation;
-    std::optional<kafka::protocol_name> protocol;
-    std::optional<kafka::member_id> leader;
-    int32_t state_timestamp;
-    std::vector<member_state> members;
-};
-
-/**
- * the key type for offset commit records.
- */
-struct group_log_offset_key {
-    kafka::group_id group;
-    model::topic topic;
-    model::partition_id partition;
-
-    bool operator==(const group_log_offset_key& other) const = default;
-
-    friend std::ostream& operator<<(std::ostream&, const group_log_offset_key&);
-};
-
-/**
- * the value type for offset commit records.
- */
-struct group_log_offset_metadata {
-    model::offset offset;
-    int32_t leader_epoch;
-    std::optional<ss::sstring> metadata;
-
-    friend std::ostream&
-    operator<<(std::ostream&, const group_log_offset_metadata&);
-};
 
 struct group_log_prepared_tx_offset {
     model::topic_partition tp;
@@ -90,29 +54,16 @@ struct group_log_aborted_tx {
 
 } // namespace kafka
 
-namespace std {
-template<>
-struct hash<kafka::group_log_offset_key> {
-    size_t operator()(const kafka::group_log_offset_key& key) const {
-        size_t h = 0;
-        boost::hash_combine(h, hash<ss::sstring>()(key.group));
-        boost::hash_combine(h, hash<ss::sstring>()(key.topic));
-        boost::hash_combine(h, hash<model::partition_id>()(key.partition));
-        return h;
-    }
-};
-} // namespace std
-
 namespace kafka {
 
 class group_stm {
 public:
     struct logged_metadata {
         model::offset log_offset;
-        group_log_offset_metadata metadata;
+        old::group_log_offset_metadata metadata;
     };
 
-    void overwrite_metadata(group_log_group_metadata&&);
+    void overwrite_metadata(old::group_log_group_metadata&&);
     void remove() {
         _offsets.clear();
         _is_loaded = false;
@@ -122,7 +73,7 @@ public:
     void update_offset(
       const model::topic_partition&,
       model::offset,
-      group_log_offset_metadata&&);
+      old::group_log_offset_metadata&&);
     void remove_offset(const model::topic_partition&);
     void update_prepared(model::offset, group_log_prepared_tx);
     void commit(model::producer_identity);
@@ -153,14 +104,14 @@ public:
         return _fence_pid_epoch;
     }
 
-    group_log_group_metadata& get_metadata() { return _metadata; }
+    old::group_log_group_metadata& get_metadata() { return _metadata; }
 
 private:
     absl::node_hash_map<model::topic_partition, logged_metadata> _offsets;
     absl::node_hash_map<model::producer_id, group::prepared_tx> _prepared_txs;
     absl::node_hash_map<model::producer_id, model::producer_epoch>
       _fence_pid_epoch;
-    group_log_group_metadata _metadata;
+    old::group_log_group_metadata _metadata;
     bool _is_loaded{false};
     bool _is_removed{false};
 };
