+++ src/v/kafka/server/tests/member_test.cc
@@ -169,7 +169,7 @@ SEASTAR_THREAD_TEST_CASE(member_serde) {
     auto m0_state = m0.state().copy();
     auto m0_iobuf = reflection::to_iobuf(std::move(m0_state));
 
-    auto m1_state = reflection::adl<kafka::member_state>{}.from(
+    auto m1_state = reflection::adl<kafka::old::member_state>{}.from(
       std::move(m0_iobuf));
     auto m1 = kafka::group_member(std::move(m1_state), m0.group_id());
 
@@ -210,7 +210,7 @@ SEASTAR_THREAD_TEST_CASE(group_log_group_metadata_compat0) {
 
     auto old_buf = reflection::to_iobuf(std::move(old));
 
-    auto new_out = reflection::from_iobuf<group_log_group_metadata>(
+    auto new_out = reflection::from_iobuf<old::group_log_group_metadata>(
       std::move(old_buf));
 
     BOOST_REQUIRE(new_out.protocol_type == "asdf");
@@ -230,17 +230,17 @@ SEASTAR_THREAD_TEST_CASE(group_log_group_metadata_compat0) {
 
 // new code + new version. will see client id/host
 SEASTAR_THREAD_TEST_CASE(group_log_group_metadata_compat1) {
-    group_log_group_metadata md;
+    old::group_log_group_metadata md;
     md.protocol_type = kafka::protocol_type("asdf");
     md.state_timestamp = 333333;
 
-    md.members.push_back(member_state{
+    md.members.push_back(old::member_state{
       .id = member_id("foobar"),
       .protocols = {{.name = protocol_name("asdf")}},
       .client_id = kafka::client_id("c0"),
       .client_host = kafka::client_host("c1"),
     });
-    md.members.push_back(member_state{
+    md.members.push_back(old::member_state{
       .id = member_id("foobar1"),
       .protocols = {{.name = protocol_name("asdf1")}},
       .client_id = kafka::client_id("c2"),
@@ -249,7 +249,7 @@ SEASTAR_THREAD_TEST_CASE(group_log_group_metadata_compat1) {
 
     auto old_buf = reflection::to_iobuf(std::move(md));
 
-    auto new_out = reflection::from_iobuf<group_log_group_metadata>(
+    auto new_out = reflection::from_iobuf<old::group_log_group_metadata>(
       std::move(old_buf));
 
     BOOST_REQUIRE(new_out.protocol_type == "asdf");
@@ -269,17 +269,17 @@ SEASTAR_THREAD_TEST_CASE(group_log_group_metadata_compat1) {
 
 // old code + new version. doesn't see client id/host
 SEASTAR_THREAD_TEST_CASE(group_log_group_metadata_compat2) {
-    group_log_group_metadata md;
+    old::group_log_group_metadata md;
     md.protocol_type = kafka::protocol_type("asdf");
     md.state_timestamp = 333333;
 
-    md.members.push_back(member_state{
+    md.members.push_back(old::member_state{
       .id = member_id("foobar"),
       .protocols = {{.name = protocol_name("asdf")}},
       .client_id = kafka::client_id("c0"),
       .client_host = kafka::client_host("c1"),
     });
-    md.members.push_back(member_state{
+    md.members.push_back(old::member_state{
       .id = member_id("foobar1"),
       .protocols = {{.name = protocol_name("asdf1")}},
       .client_id = kafka::client_id("c2"),
